Handle the UnresolvedAddressException in the createConnections function.
This exception can be thrown if the DNS for the configuration endpoint doesn't resolve. It possible that the DNS may take some time to updated and in those situations the UnresolvedAddressException will be thrown on the ch.connect(sa) call.

Currently if this exception is thrown the SocketChannel gets leaked because the sockets and never cleaned up after the exception is thrown.